+++ src/Components/Login.module.css
@@ -1,5 +1,3 @@
-
-
 .LogIn {
   background-image: linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4)),
     url('../images/login-background.jpg');
@@ -17,3 +15,5 @@
   text-align: center,
 
 }
+
+
