+++ src/Components/User/TripListItem.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useDocument } from 'react-firebase-hooks/firestore';
-import db from '../firebase';
+import db from '../../firebase';
 import { Link } from 'react-router-dom';
 import { ListGroup, Spinner, CardGroup } from 'react-bootstrap';
-import TripResultCard from './TripResultCard';
+import TripResultCard from '../Helper/TripResultCard';
 
 export const TripListItem = props => {
   const [value, loading, error] = useDocument(db.doc(`Trips/${props.tripId}`), {
