+++ src/Components/index.js
@@ -1,26 +1,26 @@
 /* eslint-disable no-unused-vars */
-import TripSingle from './TripSingle';
-import UserProfile from './UserProfile';
-import UserProfileHeader from './UserProfileHeader';
-import UserProfileTrips from './UserProfileHeader';
+import TripSingle from './SingleTrip/TripSingle';
+import UserProfile from './User/UserProfile';
+import UserProfileHeader from './User/UserProfileHeader';
+import UserProfileTrips from './User/UserProfileHeader';
 import Login from './Login';
-import SearchAPI from './SearchAPI';
-import TripPlanning from './TripPlanning';
-import TripSearch from './TripSearch';
-import Signup from './Signup';
-import BucketListCard from './BucketListCard';
-import BucketList from './BucketList';
-import WorldMap from './WorldMap'
+import SearchAPI from './TripPlanning/SearchAPI';
+import TripPlanning from './TripPlanning/TripPlanning';
+import TripSearch from './TripPlanning/TripSearch';
+import BucketListCard from './User/BucketListCard';
+import BucketList from './User/BucketList';
+import WorldMap from './User/WorldMap'
+import CreateTrekkForm from './SingleTrip/CreateTrekkForm'
 
-export { default as TripSingle } from './TripSingle';
-export { default as UserProfile } from './UserProfile';
-export { default as UserProfileHeader } from './UserProfileHeader';
-export { default as UserProfileTrips } from './UserProfileTrips';
+export { default as TripSingle } from './SingleTrip/TripSingle';
+export { default as UserProfile } from './User/UserProfile';
+export { default as UserProfileHeader } from './User/UserProfileHeader';
+export { default as UserProfileTrips } from './User/UserProfileTrips';
 export { default as Login } from './Login';
-export { default as SearchAPI } from './SearchAPI';
-export { default as TripPlanning } from './TripPlanning';
-export { default as TripSearch } from './TripSearch';
-export { default as Signup } from './Signup';
-export { default as BucketListCard } from './BucketListCard';
-export { default as BucketList } from './BucketList';
-export { default as WorldMap } from './WorldMap';
+export { default as SearchAPI } from './TripPlanning/SearchAPI';
+export { default as TripPlanning } from './TripPlanning/TripPlanning';
+export { default as TripSearch } from './TripPlanning/TripSearch';
+export { default as BucketListCard } from './User/BucketListCard';
+export { default as BucketList } from './User/BucketList';
+export { default as WorldMap } from './User/WorldMap';
+export { default as CreateTrekkForm } from './SingleTrip/CreateTrekkForm';
