+++ src/Routes.js
@@ -7,11 +7,10 @@ import {
   SearchAPI,
   TripPlanning,
   TripSearch,
-  Signup,
   WorldMap,
 } from './Components';
-import {PhotoLoad} from './Components/PhotoLoad'
-import {Welcome} from './Components/Welcome'
+import {PhotoLoad} from './Components/Helper/PhotoLoad'
+import {Welcome} from './Components/Welcome/Welcome'
 
 export class Routes extends Component {
   render() {
@@ -21,7 +20,6 @@ export class Routes extends Component {
         <Route exact path="/plantrip" component={TripPlanning} />
         <Route path="/trip/:tripId" component={TripSingle} />
         <Route path="/login" component={Login} />
-        <Route path="/signup" component={Signup} />
         {/* Testing Routes */}
         <Route path="/triposoAPI" component={SearchAPI} />
         <Route exact path="/tripSearch" component={TripSearch} />
